Change ising_hamiltonian function to only use l
There is a bug in the ising_hamiltonian function. It works in the tutorial as long as you use a range in the training phase of 2 but when you changed this to 3 or more it would raise an exception. I believe this is due to the fact that the training epoch iterator is "i" and for some reason the ising_hamiltonian function also used "i" even though I don't believe that is should (or has to). The q_weights passed to it are size (4,3) so I believe the better way to do this would be to use "RY(weights[l, l-1], wires=w)"