More consistent and cleaner variable naming
`q_net` and `Quantumnet` are very confusing names for two very different
things. When first approaching this tutorial it takes some time to
understand what's going on and I tried to provide some more clarity by
renaming them to `quantum_net` and `DressedQuantumNet`. These names are more in
line with their descriptions in the text.