[Demo] Perturbative Gadgets for Variational Quantum Algorithms (#610)
* Adding a demo on perturbative gadgets for VQE

Adding the tutorial script (converted from jupyter) and dependencies (in ./barren_gadgets/) necessary for the tutorial on perturbative gadgets for VQE

* gallery link for the barren gadgets demo

Editing the demos_optimization.rst file to add a gallery link and an entry in the table of contents

* Delete barren_gadgets/__pycache__ directory

This directory was created directly when including the repository with the auxiliary files for the tutorial, not sure why

* Moving the demo files in the right folder

* Update demonstrations/tutorial_barren_gadgets.py

* Apply suggestions from code review - references

Changing the references format so that they automatically link with the bibliography

Co-authored-by: Josh Izaac <josh146@gmail.com>

* changing links to other demos

Using local links to other demos instead of their URL s.t. the linking is not affected by a change of URL of the website

Co-authored-by: Josh Izaac <josh146@gmail.com>

* Fix references

Updating the first paper reference to see if the automatic linking works

* Plots: imports and style

Changing the location of the library imports (for didactic purposes) and the styles of the plots, following @KetpuntoG's suggestions

* Description of custom functions

Added an explanation of the imported (custom) functions, and links to the files

* Add placeholder for author bio

Co-authored-by: Josh Izaac <josh146@gmail.com>

* author picture

Adding a picture for the author bio

* Cleaning up

Removing some functions that are not used in the tutorial

* Explanations

Adding a final paragraph on where to find more information, and fixing a few aesthetics details

* Update demonstrations/tutorial_barren_gadgets.py

* Update demonstrations/tutorial_barren_gadgets.py

* Apply suggestions from code review

Accepting the suggestions. Mainly small reformulations and spotting typos + some clarifications

Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>

* Minor fixes

Solving details in punctuation and file links

* Changing Hamiltonian notation

Switching from qml.Hamiltonian(coeffs, obs) to using the `@` notation

* Hamiltonian mistake

I had changed the Hamiltonian compared to previous versions of the demo, now it's the same as it was until before the notation change.

* Update demonstrations/tutorial_barren_gadgets.py

* Update demonstrations/tutorial_barren_gadgets.py

* Definition: barren plateaus

* Definitions: global and local

* Edit conclusion

* typos and notation

Few minor adjustments and being consistent in the naming of the target/computational Hamiltonian. Now only using target

* More target vs computational

* Qubit excess

The training example used more auxiliary qubits that needed (two 4-qubit registers instead of one). Removed those qubits and looked for a new set of initial conditions that gives a nice visualization. Also changed the trained Hamiltonian to be the first term of the 2-term example given above (don't know why they were different)

* global vs local

Adding some more emphasis on the local and global example Hamiltonians being different Hamiltonians but with the same ground state

* Update author bio

* Removing prints

Removing the prints from the training

* Minor changes

Polishing some equations, adding some links, slight reformulations and variable renaming

* Updating image

Updating the image to use the 'i' index instead of 's'

* Move matplotlib import

Moving the maptplotlib import and adding a show statement to avoid the output print

* Apply suggestions from code review (links)

updating some of the links used in the demo

Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>

* Update links

Changing the 'here' from the link to the repo (which apparently does not work in some browsers)

* Remove unnecessary line breaks

* Author bio link

Changing the link in the author bio to be stated in plane text

* Apply suggestions from code review

Adopting the typo corrections and small reformulations proposed

Co-authored-by: Catalina Albornoz <albornoz.catalina@hotmail.com>

* Update demonstrations/tutorial_barren_gadgets.py

* Apply suggestions from code review

Co-authored-by: Catalina Albornoz <albornoz.catalina@hotmail.com>

* Author bio

Updating the bio to be more precise about the "title"

* Apply suggestions from code review

Co-authored-by: Catalina Albornoz <albornoz.catalina@hotmail.com>

* Auto stash before merge of "master" and "origin/master"

* Revert "Auto stash before merge of "master" and "origin/master""

This reverts commit 509c0a5552bbc97f55abf57b4f4bb33fe77c034b.

* Minor rewording

* Apply suggestions from code review

Accepting most changes proposed by @isaacdevlugt

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>

* Rephrasing + hook

Proposing a version of the resolution/precision discussion and adding to the introduction

* final changes

* updating cells

* Update demonstrations/tutorial_barren_gadgets.py

Co-authored-by: Catalina Albornoz <albornoz.catalina@hotmail.com>

* weird spaces

* Update tutorial_barren_gadgets.py

* Update demonstrations/tutorial_barren_gadgets.py

* hamiltonians

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>
Co-authored-by: Catalina Albornoz <albornoz.catalina@hotmail.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>