Initial state preparation demo (#950)
**Title:**

Initial state preparation for quantum chemistry

**Summary:**

This demo is explaining the concept of an initial state -- a primer for
a quantum algorithm that improves its chances of success specifically
for the task of obtaining the ground-state energy of a molecule. It
showcases the newly built functionality in PennyLane for importing
outputs from sophisticated computational chemistry methods to initialize
better initial states. Finally, it shows how the choice of initial state
affects the execution of VQE on a simple molecule, demonstrating the
importance of initial state for quantum algorithms and that this can of
algorithms R&D can be now performed in PennyLane.

**Relevant references:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: soranjh <soran.jahangiri@gmail.com>
Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>
Co-authored-by: Alvaro Ballon <91897656+alvaro-at-xanadu@users.noreply.github.com>
Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>