Minor fixes for version v0.33 (#963)
In version v0.33, we ported transforms over to a new syntax. This
requires a few syntax changes in the demo.

We also switched from the old default qubit to the new one. The old
device used the global random number generator each time it sampled. The
new device will query the global random number generator once upon
initialization if a seed is not specified. This caused the global random
number generator to be in a different state when creating parameters.

In order to avoid coupling between random number generation in different
parts of the workflow in the future, I switched to using local random
number generation for parameters and setting a seed explicitly on the
device. Future changes to random generation shouldn't have as much
impact on the output of the demos.

A couple of deprecation warnings from matplotlib and torch also forced a
couple of updates.