Bump Python version to 3.10. Bump Poetry version to 1.7.1 (#996)
**Title:** Bump Python Build version of QML to Python 3.10

**Summary:**
This PR drops Python 3.9 from QML and upgrades all CI to use 3.10 for
building. This ensures compatibility with more packages including
pennylane itself.

**Relevant references:**

[sc-51655](https://app.shortcut.com/xanaduai/story/51655/attempt-to-build-qml-using-python-3-10)

**Possible Drawbacks:**
There could potentially be rendering issues with react as the version of
sphinx has been bumped from 4.0 to 4.3 (that is the lowest version of
sphinx compatible with Python 3.10).

**Related GitHub Issues:**
None.