update auto-run demos before 0.34 release (#1005)
- learning2learn was failing because we call a QNode with a vanilla list
of parameters at some point, and it wrongly auto-assumed that the
interface was `autograd`. Not sure why this only came up now, but
forcing it to `tf` fixes it
- the Barrier in the teleportation demo was just to make it look like
the two measurements happen simultaneously, but that doesn't happen
anymore regardless... so I removed it
- the second full circuit MPL drawing (with the QNode) is supposed to
show that defer_measurements is applied, so I had to add
`expansion_strategy="device"` to the QNode to make sure it is applied
(it used to be applied at qnode-level, but now it's part of device
pre-processing)
- question: should I put the kwarg in the qnode construction, or in the
invocation of `draw_mpl`? I think they both accept it
- unitary_designs was using the old `qfunc_transform` and a warning was
being raised. Updated to the new `transform` API to avoid the warning