Update hybrid jobs demos (#1009)
The hybrid jobs were failing. Some of them were because of incorrect
variable names. Others because functions defined outside the hybrid job
context were being used inside it. This PR fixes both of those.

From discussions outside GitHub:

> * The `getting_started_with_hybrid_jobs` demo uses some pickling and
is only expected to work with 0.33.1. Once changes in this PR are
incorporated, it does work with 0.33.1. An empty dictionary for results
is returned with the most recent PL, and it's not a bug, its built in on
the Braket end.
> * Once the fixes for `braket_parallel_gradients` are added, everything
runs, the results are just incorrect. It's a static demo, so it will
continue to look like the results are correct unless someone actually
tries it. We'll let AWS know.

There is also a deprecation warning about `Observable.return_types` in
the `braket_parallel_gradients` demo, but it's originating from the
braket plugin. Their release is incosistent with ours, so it will be
fixed, but probably not in time for 0.34

---------

Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>