new thumbnails (#1014)
Adding the new thumbnails. I have chosen to create a folder with the
thumbnails separately (following the idea of ​​large thumbnails).

In the end I chose not to delete the old ones. They remain in their
original demo folder. This is because they could be used within the demo
for explanation purposes. **This PR does not create any inconsistency
with the other repositories** , I'll update them to the new images

---------

Co-authored-by: Alvaro Ballon <91897656+alvaro-at-xanadu@users.noreply.github.com>