Add how-to's on JAX optimization (#970)
**Context**

The PennyLane demos have a lot of 'explanation' content -- good for when
you are learning, and for understanding new theory and its context
(alongside code). However, we are lacking in 'how to' content.

How-to content is a bit different from a tutorial (teaching someone the
basics) or a demo (teaching concepts and understanding). If you think of
a tutorial as a 'Basic cooking lesson', and a demo as a 'Book on the
history of cooking', you can think of a how-to as a **recipe**.

A how-to teaches the reader practical knowledge, but is not presented as
a lesson. Instead, it is content that is useful when working, and you
need to remember how to do something.

Rather than teaching the basic concepts like a tutorial does, how-to's
will:

- Assume basic understanding and familiarity.
- Provide concrete steps.
- Discuss options and practical alternates where appropriate, because it
won’t often be clear what the users real-world situation is.

For these two how-tos, I tried to follow the following guides:

- **Do not teach.** Just like recipes don’t attempt to teach readers the
basic cooking skills, how-to’s shouldn’t either. Instead, link out to
tutorials and reference documentation if needed.
- Provide concrete steps, and assume familiarity with basic concepts and
features
- Be practical. Don’t get distracted by explaining theoretical or
abstract concepts.
- Be concise. Use simple, straightforward language.
- Be consistent. Make sure that examples provided follow on and are
consistent throughout all steps.

**Details**

The two how-to's are available here:

- [How to optimize a quantum machine learning model using JAX and
Optax](https://qml-build-previews.pennylane.ai/pull_request_build_preview/970/demos/tutorial_How_to_optimize_QML_model_using_JAX_and_Optax.html)

- [How to optimize a quantum machine learning model using JAX and
JAXopt](https://qml-build-previews.pennylane.ai/pull_request_build_preview/970/demos/tutorial_How_to_optimize_QML_model_using_JAX_and_JAXopt.html)

**Drawbacks**

For now, I have categorised these simply under existing categories.
However:

- I have added a 'how to' tag.
- In the future, it might be nice to easily show all demos tagged as
how-to.

---------

Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>
Co-authored-by: Maria Schuld <mariaschuld@gmail.com>