Learning shallow quantum circuits (#1023)
First draft of a quick turnover demo on
[2401.10095](https://arxiv.org/abs/2401.10095)

It mainly shows the sewing procedure and one possible way of obtaining
local inversions (which is not really discussed in the paper
unfortunately)

Still a draft. In particular, the numerical example part at the end
still needs text. Just happy that it works now for the moment.

---------

Co-authored-by: Maria Schuld <mariaschuld@gmail.com>
Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>