Demo RL with pulse programming (#985)
**Title:** Reinforcement learning for gate calibration

**Summary:** In this demo we learn how quantum gates are implemented in
superconducting quantum circuits in practice. We will understand why we
need to calibrate every qubit individually and how to deal with this
process in an experimentally-friendly way with reinforcement learning.

**Relevant references:** It implements a protocol similar to the one
introduced in [this
paper](https://journals.aps.org/prxquantum/abstract/10.1103/PRXQuantum.2.040324)

----
If you are writing a demonstration, please answer these questions to
facilitate the marketing process.

* GOALS — Why are we working on this now?
It is related to the project I developed during my summer residency.
It's a cool application of the pulse module in combination with
reinforcement learning that can be easily taken from simulations to
actual real devices.

* AUDIENCE — Who is this for?
For anyone in the field of quantum computing looking for an excuse to
learn about reinforcement learning.

* KEYWORDS — What words should be included in the marketing post?
reinforcement learning, pulse programming, ... (need to think about
these)

* Which of the following types of documentation is most similar to your
file?
I currently don't have access to the linked notion page. It's on the
lengthy side.

---------

Co-authored-by: Tom Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>