[MCM 1/3] How-to: Statistics of mid-circuit measurements (#1066)
**Title:**
How to collect statistics of mid-circuit measurements

**Summary:**
PennyLane supports mid-circuit measurements not only to create dynamic
circuits with control flow using the measurement results, but also
supports a wide range of postprocessing options of the measured values.
This how-to explains the usage of `qml.measure` and briefly compares
`qml.defer_measurements` and `qml.dynamic_one_shot`. Then it shows how
(not) to measure qubit correlations with MCMs, and afterwards walks
through the supported postprocessing of MCM values and the allowed
return functions such as `qml.expval` or `qml.counts`.
Finally, ____ is shown as an application for obtaining statistics of
MCMs.

**Relevant references:**
N/A

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
N/A

----
If you are writing a demonstration, please answer these questions to
facilitate the marketing process.

* GOALS — Why are we working on this now?

This how-to promotes mid-circuit measurement functionality in PennyLane,
which has been extended in recent releases. In particular, it contains a
comparison of the brand-new native MCM support via `dynamic_one_shot`
with `defer_measurements` and can serve to advertise this new feature.

In addition, this how-to is part of the broader goal of emphasizing MCMs
and dynamic circuits in our contents.

* AUDIENCE — Who is this for?

Typical how-to audience: Fundamental concepts of measurements and
quantum computing are known, but the audience may not yet be familiar
with PennyLane's syntax for MCMs and returning their outcomes, or they
may want to consult this how-to as a reference of how to realize their
circuit with MCMs.
Anyone really with the goal of coding up a circuit that returns MCM
values can be in the audience.

* KEYWORDS — What words should be included in the marketing post?

  - Mid-circuit measurements
  - Statistics
  - [optional] postprocessing


* Which of the following types of documentation is most similar to your
file?
(more details
[here](https://www.notion.so/xanaduai/Different-kinds-of-documentation-69200645fe59442991c71f9e7d8a77f8))
    
- [ ] Tutorial
- [ ] Demo
- [x] How-to

[sc-60836]

---------

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>