[WIP] [g-sim 3] (g + P)-sim (#1118)
Merged the approach from https://github.com/PennyLaneAI/qml/pull/1106
with the branching approach here. This has neater illustrations anyway 👍

If you are writing a demonstration, please answer these questions to
facilitate the marketing process.

* GOALS — Why are we working on this now?
Lie algebras in QC and simulation techniques like g-sim are a hot topic
due to their relevance to barren plateaus. This is an extension of g-sim
to a small number of non-DLA gates. It bears some intriguing theoretical
features but, like g-sim itself, is of rather limited practical
relevance.

* AUDIENCE — Who is this for?
This is an expert demo targeted towards people familiar with Lie
algebras, g-sim in particular.

* KEYWORDS — What words should be included in the marketing post?
  - dynamical Lie algebra / Lie groups
  - g-sim

* Which of the following types of documentation is most similar to your
file?
(more details
[here](https://www.notion.so/xanaduai/Different-kinds-of-documentation-69200645fe59442991c71f9e7d8a77f8))
    
- [ ] Tutorial
- [x] Demo
- [ ] How-to 

[sc-62859]

---------

Co-authored-by: Ivana Kurečić <ivana@xanadu.ai>