Quantum volume updates (#1146)
The lima device used in the quantum volume demo no longer exists. This
update switches to make it use the fake lima device instead.

Since when running the demo, I got data that no longer agreed with the
conclusions of the data present in the demo, I added a warning that
device details may have changed over time.

Since we now use a fake device instead of a real device, this demo could
theoretically be automatically run. But the difference in the data
output means that the automatically run results would not tell the same
story.

---------

Co-authored-by: Olivia Di Matteo <2068515+glassnotes@users.noreply.github.com>