Qubitization demo (#1113)
**Title:** I

ntro to Qubitization

**Summary:**

Encoding a Hamiltonian into a quantum computer is a fundamental task for
many applications, but the way to do it is not unique. One method that
has gained special status is known as Qubitization. In this demo, we
will introduce the Qubitization operator and explain how to view it as a
rotation operator.

**Relevant references:**

**Possible Drawbacks:**

**Related GitHub Issues:**

----
If you are writing a demonstration, please answer these questions to
facilitate the marketing process.

* GOALS — Why are we working on this now?

Support the content of the Algorithm Toolkit. `qml.Qubitization` has
been one of the latests additions

* AUDIENCE — Who is this for?

everyone, it should be easy to understand

* KEYWORDS — What words should be included in the marketing post?


In this demo we propose a cool way to visualize the algorithm. It could
be good to refer to that

Qubitization, block encoding, visualization, algorithm toolkit

* Which of the following types of documentation is most similar to your
file?
(more details
[here](https://www.notion.so/xanaduai/Different-kinds-of-documentation-69200645fe59442991c71f9e7d8a77f8))
    
- [ ] Tutorial
- [x ] Demo
- [ ] How-to

---------

Co-authored-by: ixfoduap <40441298+ixfoduap@users.noreply.github.com>
Co-authored-by: Ivana Kurečić <ivana@xanadu.ai>