Update workflows to upload and sync objects.inv to dev, staging and prod environments (#1216)
**Title:** Update workflows to upload and sync objects.inv to dev,
staging and prod environments

**Summary:** 
Currently we do not sync the generated objects.inv file from QML sphinx
build. This causes issues when building docs for pennylane docs as
linking to things within QML fails.

A quick solution was being used currently where the objects.inv was
being manually updated, but this PR changes the workflow so that it is
done with merges to dev/master branches,

**Relevant references:**

[sc-73313](https://app.shortcut.com/xanaduai/story/73313/update-qml-pipeline-to-save-objects-inv-from-each-master-dev-branch-build)


**Possible Drawbacks:**
None.

**Related GitHub Issues:**
None.

----
If you are writing a demonstration, please answer these questions to
facilitate the marketing process.

* GOALS — Why are we working on this now?

*Eg. Promote a new PL feature or show a PL implementation of a recent
paper.*

N/A

* AUDIENCE — Who is this for?

*Eg. Chemistry researchers, PL educators, beginners in quantum
computing.*

N/A. Internal facing changes.

* KEYWORDS — What words should be included in the marketing post?


* Which of the following types of documentation is most similar to your
file?
(more details
[here](https://www.notion.so/xanaduai/Different-kinds-of-documentation-69200645fe59442991c71f9e7d8a77f8))
    
- [ ] Tutorial
- [ ] Demo
- [ ] How-to