Combine benchmarks for first and subsequent QJIT calls
The benchmarks for the first and subsequent calls to the QJIT-compiled
circuit were initially kept separate since the first call incurred some
additional, then unknown, overhead. This overhead is now understood to
be due to loading and initializing some JAX and XLA utilities. Since we
now call the QJIT-compiled circuit earlier in the demo, this loading/
initializing occurs there, so we no longer measure any difference in
runtime of the first and subsequent calls to the QJIT circuit when we
execute them in the benchmarking section.

Also reword the text of the demonstration to account for this change.