Demo on " Constant-depth preparation of matrix product states with dynamic circuits" (#1185)
**Title:**
Constant-depth preparation of matrix product states with dynamic
circuits

**Summary:**
Matrix product states (MPS) form an important class of quantum many-body
states. Preparing them is an important subroutine for any algorithm that
want to use MPS as an initial state, for example. A recent paper, on
which this demo is based, and which it is following closely, generalized
a method that reduces the depth of such a preparation circuit to
constant scaling, by using auxiliary qubits and dynamic quantum circuits
(mid-circuit measurements together with classical feedforward).
We here focus on the simplest scenario in the main text of the paper,
and implement a simple example in code.
The implemented MPS has a parameter controlling the correlation length
of the prepared state, which we verify explicitly.

**Relevant references:**
https://arxiv.org/pdf/2404.16083
https://arxiv.org/pdf/quant-ph/0501096
https://journals.aps.org/prxquantum/pdf/10.1103/PRXQuantum.4.020315

**Possible Drawbacks:**

**Related GitHub Issues:**

----
If you are writing a demonstration, please answer these questions to
facilitate the marketing process.

* GOALS — Why are we working on this now?

Mid-circuit measurements recently are/were a focus point in PennyLane.
MPS states are used in some quantum chemistry algorithms as initial
state.


* AUDIENCE — Who is this for?

  Researchers in quantum algorithms


* KEYWORDS — What words should be included in the marketing post?

TBD

* Which of the following types of documentation is most similar to your
file?
(more details
[here](https://www.notion.so/xanaduai/Different-kinds-of-documentation-69200645fe59442991c71f9e7d8a77f8))
    
- [ ] Tutorial
- [x] Demo
- [ ] How-to

[sc-64275]

---------

Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Ivana Kurečić <ivana@xanadu.ai>