Update `update-dev.yml` to use `git` instead of third-party action (#1242)
The `update-dev` action keeps failing due to unrelated histories between
the `dev` and `master` branches, and the third party action used for the
nightly merge does not include any options to allow merging with
unrelated histories. So, this PR updates the action to use `git`
instead.

This will also give us more flexibility in the future.

Couple of updates:
* I have confirmed that configuring `git` using `secrets.NIGHTLY_TOKEN`
does not expose the token in the workflow logs. See
[here](https://github.com/PennyLaneAI/qml/actions/runs/11408914907/job/31748090296#step:3:4)
for an example.

[sc-76254]

---------

Co-authored-by: Rashid N H M <95639609+rashidnhm@users.noreply.github.com>