Add author profiles (#1258)
Apologies for the large PR, but this touches a large number of demos.

This PR replaces legacy demo author information with their new pennylane
Profiles. It also introduces a new version of the `demo.metadata.schema`
which requires the `username` field (PennyLane Profile), and will not
accept the legacy `id` field for demo authors. The expectation is that
all demo authors create PennyLane Profiles to store their avatar and bio
going forward.

*Notes:*
The new metadata schema file is exactly the same as the previous version
(copy paste) except for the `authors` field:
```
      "authors": {
        "description": "The author(s) of the demo. This array must contain at least one item.",
        "type": "array",
        "items": {
          "$ref": "file:./objects/author.schema.0.2.0.json"
        },
        "minItems": 1
      },
```

---------

Co-authored-by: GitHub Nightly Merge Action <actions@github.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Mikhail Andrenkov <mikhail@xanadu.ai>
Co-authored-by: David Wierichs <davidwierichs@gmail.com>
Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Ivana Kurečić <ivana@xanadu.ai>
Co-authored-by: Jack Brown <jack.brown486@gmail.com>
Co-authored-by: bellekaplan <167111433+bellekaplan@users.noreply.github.com>
Co-authored-by: Utkarsh <utkarshazad98@gmail.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: Guillermo Alonso-Linaje <65235481+KetpuntoG@users.noreply.github.com>
Co-authored-by: gmlej <gmlejarza@gmail.com>
Co-authored-by: Jorge J. Martínez de Lejarza <61199780+gmlejarza@users.noreply.github.com>
Co-authored-by: ixfoduap <40441298+ixfoduap@users.noreply.github.com>