QA updates v0.40 lillian (#1296)
- Some of the `ops` print differently after switching from `Tensor` to
`Prod`. Cleaned that up.
- The link to the PennyLane-Qiskit device documentation in the Ensemble
PR is a dead link to a device we no longer have (`qiskit.ibmq`). In the
new plugin, we have `qiskit.remote`.

The second point could potentially be resolve differently. The device
name _is_ correct for the version of the plugin that the demo works
with, but the remote device should also be fine with that version. We
could also leave it mentioning `qiskit.ibmq` because it's an old demo,
and just delete the link 🤷‍♀️ But we don't want dead links.