Update `tutorial_noisy_circuits.py` to port `default.mixed` free of legacy interfaces (#1284)
**Title:**
Update `tutorial_noisy_circuits.py` to port `default.mixed` free of
legacy interfaces

**Summary:**
In the Tutorial of Noisy Circuits, there's a call `dev.state` which is a
legacy-only interface. In our current standard, the device object will
not store the state any more and we should return the latest state via
the corresponding measure.

**Relevant references:**

**Possible Drawbacks:**

**Related GitHub Issues:**

----

---------

Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>