Add react master token (#1340)
The PAT tokens for pulling the PL REACT site keep getting overwritten by
Terraform. These will be redundant soon as we switch to the new QML
pipeline, so for now I am swapping these all out with a master token to
keep builds moving. This token has already been provisioned and
verified. It will expire in June, by which time the new pipeline should
be ready.