[Bug fix] Update fixed depth hamiltonian simulation demo (#1344)
The problem is that the computed horizontal CSA $\mathfrak{h}$ is not
necessarily composed of just Pauli words, but soms of Paulis (Pauli
sentence). The old version was assuming everything was Pauli, did the
decomposition in terms of Pauli words, and accordingly counted elements.

Updating this here to decompose in the actual basis of the horizontal
CSA $\mathfrak{h}$ and show explicitly that the decomposition is stricly
in it.

edit: this works all well locally with correct jax, optax and pl
version, but for some reason fails in CI 🅰️ 🅰️ 🅰️
edit2: increasing the tolerance in `horizontal_cartan_subalgebra` fixes
this. A bit problematic that this cannot be reproduced locally 😢

---------

Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: David Wierichs <davidwierichs@gmail.com>