Update PL and plugin dependencies (#1350)
This PR targets `dev` and it updates the dependencies to use the new
PennyLane version, as well as the new plugin versions.

We need to use `poetry` to update the dependencies.

Finally, we also revert the temporary change of PennyLane dependency in
the Makefile at the beginning of feature freeze.