Decrease MBQC demo size (#1382)
The MBQC demo is ~15MB, which is too large to be uploaded to our V2 demo
backend (10MB max). This is also a bad experience for any users with a
slower internet connection, as this demo page will take a long time to
load. The biggest contributor is the .gif, which is ~14MB. I've cut it
down in time so it's shorter and under 5MB.